==> pipeline/1-get-data.ipynb <==
## School Performance

### 2023 3rd Grade Reading Performance: % of students achieving ‘meets’ on the STAAR assessment (state exam):

Visited: [Texas Agency Performance Reporting](https://rptsvr1.tea.texas.gov/cgi/sas/broker?_service=marykay&year4=2023&year2=YEAR2&prgopt=2023/xplore/pickcol.sas&_program=perfrept.perfmast.sas&_debug=0&step=1+&steps=2&bylev=&format=XLS&camp0=999999&sumlev=C&dsname=CSTAAR_GR3)

Downloaded: "Grade 3, Reading, Meets Grade level, 2023" and "Campus Name, District Name"

Output: data/raw/CSTAAR_GR3.csv

### Reference File

Data for "Campus 2023 Domain 1a: Meets Grade Level Std, Grade 3, All Students, Staar Reading/Ela Rate" is found in column CDA03ARE1223R

### Data Dictionary

Source: https://rptsvr1.tea.texas.gov/perfreport/tapr/2023/datadict.pdf

Outputs: data/raw/cstaar_datadict.pdf

## Community Data / Indicators

### School Addresses

Visited: https://strivetogether.egnyte.com/dl/1DmPUCs1Nr

Downloaded:  "School Addresses Hiring Project"

Output: data/raw/School Addresses - Hiring Project.xlsx

### Median Household Income by Zip Code

Source: Census Explorer: Table [S1901](https://data.census.gov/table/ACSST5Y2022.S1901?q=household%20income&g=040XX00US48$8600000&y=2022): ACS Income in past 12 months

Metric: Median income (dollars)

Year: 2022 5-year estimates

Output: 
* data/raw/ACSST5Y2022.S1901_2024-12-22T150557.zip
* data/raw/ACSST5Y2022.S1901_2024-12-22T150557/ACSST5Y2022.S1901-Column-Metadata.csv
* data/raw/ACSST5Y2022.S1901_2024-12-22T150557/ACSST5Y2022.S1901-Data.csv
* data/raw/ACSST5Y2022.S1901_2024-12-22T150557/ACSST5Y2022.S1901-Table Notes.txt

### Poverty Data by Zip Code

Source: Census Explorer: Table [S1701](https://data.census.gov/table/ACSST5Y2022.S1701?q=povety&g=040XX00US48$8600000&y=2022): Poverty Status in Past 12 Months

Metric: Percent Below Poverty Level

Year: 2022 5-year estimates

Outputs:
* data/raw/ACSST5Y2022.S1701_2024-12-22T150853.zip
* data/raw/ACSST5Y2022.S1701_2024-12-22T150853/ACSST5Y2022.S1701-Column-Metadata.csv
* data/raw/ACSST5Y2022.S1701_2024-12-22T150853/ACSST5Y2022.S1701-Data.csv
* data/raw/ACSST5Y2022.S1701_2024-12-22T150853/ACSST5Y2022.S1701-Table-Notes.txt
