==> pipeline/4-agg-zipcode.ipynb <==
import pandas as pd

# Inputs

reading = pd.read_csv('../data/interim/school_scores.csv')

# Agg

# These are id columns that we can drop:

reading = reading.drop(columns=['campus_number','city','state','_merge','zip_plusfour'])

# These are fixed effects we can summarize. XXX: Consider doing weighted mode (by enrollment)?
# TODO: Going to ignore these for now.
fixed_effects = {
    "instruction_type": lambda x: x.value_counts(dropna=False).idxmax(),
    "charter_type": lambda x: x.value_counts(dropna=False).idxmax(),
    "grade_range": lambda x: x.value_counts(dropna=False).idxmax(),
    "status": lambda x: x.value_counts(dropna=False).idxmax(),
}

reading_by_zip = reading.groupby('zip').agg({
    "reading": ['min','mean','max','std'],
    "enrollment": "sum",
    'district_name': ["nunique","unique"],
    'campus_name': "nunique",
}).reset_index()

reading_by_zip.columns = reading_by_zip.columns.map('_'.join).map(lambda x: x.strip('_'))
reading_by_zip = reading_by_zip.rename(columns={"enrollment_sum":"total_enrollment"})

# Export

reading_by_zip.to_csv("../data/interim/zipcode_scores.csv", index=False)
