==> pipeline/7-get-geoms.ipynb <==
import pygris
import pandas as pd
import geopandas as gpd

from us import states

geoms = pygris.zctas(cb=True, year=2020)

tx_zips = gpd.read_file(states.TX.shapefile_urls()['zcta'])

assert all(geoms['ZCTA5CE20'] == geoms['NAME20'])
assert all(geoms['ZCTA5CE20'] == geoms['GEOID20'])

geoms = geoms.rename(columns={"ZCTA5CE20":"zip"}).filter(['zip','geometry'])

geoms = geoms[geoms.zip.isin(tx_zips.ZCTA5CE10)]

geoms.to_file("../data/final/zipcodes.geojson", index=False)
